refactor: Separate config files in config directory
