+++ .eslintrc.js
@@ -11,5 +11,7 @@ module.exports = {
         'linebreak-style': 0,
         'no-plusplus': 0,
         indent: ['error', 4],
+        'max-len': 0,
+        'no-unused-vars': ['error', { args: 'none' }],
     },
 };
